lib: Support csr detect suuport for Zcheripurecap ABI
- Properly handle capability pointer in csr_read_allowed() and
  csr_write_allowed().
- Return capability pointer of trap handler from
  sbi_hart_expected_trap_addr().

+++ include/sbi/sbi_hart.h
@@ -111,9 +111,9 @@ int sbi_hart_reinit(struct sbi_scratch *scratch);
 int sbi_hart_init(struct sbi_scratch *scratch, bool cold_boot);
 
 extern void (*sbi_hart_expected_trap)(void);
-static inline ulong sbi_hart_expected_trap_addr(void)
+static inline uintptr_t sbi_hart_expected_trap_addr(void)
 {
-	return (ulong)sbi_hart_expected_trap;
+	return (uintptr_t)sbi_hart_expected_trap;
 }
 
 unsigned int sbi_hart_mhpm_mask(struct sbi_scratch *scratch);
