+++ lib/utils/irqchip/imsic.c
@@ -241,7 +241,7 @@ void imsic_local_irqchip_init(void)
 	 */
 
 	/* If Smaia not available then do nothing */
-	csr_read_allowed(CSR_MTOPI, (ulong)&trap);
+	csr_read_allowed(CSR_MTOPI, (uintptr_t)&trap);
 	if (trap.cause)
 		return;
